refactor: 메시지 관리 통합으로 인한 클래스명 변경
