+++ src/main/java/com/codenear/butterfly/auth/application/email/EmailLoginService.java
@@ -1,0 +1,4 @@
+package com.codenear.butterfly.auth.application.email;
+
+public class EmailLoginService {
+}
